Integrate cats-time into cats-kernel.
The `cats-time` microlibrary publishes instances of the typeclasses of
cats.kernel and cats for the classes in the java.time package.
In [Issue 3766](https://github.com/typelevel/cats/issues/3766) it was
proposed to integrate this library into cats itself, considering that
for JVM users the `java.time` classes are just as standard as the
UUID or the Scala Standard Library ones.